+++ src/data/layout/queue/destaque/methods/mod.rs
@@ -1,7 +1,0 @@
-// devela/src/data/layout/queue/destaque/methods/mod.rs
-//
-//! Double-ended queue + double-ended stack.
-//
-
-mod general;
-// mod own; // TODO