+++ src/data/layout/stack/stack/methods/mod.rs
@@ -1,8 +1,0 @@
-// devela/src/data/layout/stack/stack/methods/mod.rs
-//
-//! Stacks.
-//
-
-mod convert;
-mod general;
-mod own;