+++ docs/changelog.md
@@ -20,7 +20,7 @@
 
 ## features & flags
 - new features: `code_all`, `data_all`, `ffi`, `geom_all`, `lang_all`, `media_all`, `num_all`, `org_all`, `phys_all`, `run_all`, `sys_all`, `text_all`, `ui_all`, `vita_all`, `widget`, `work_all`.
-- remove features: `layout`, `safe_color`, `safe_draw`, `safe_image`.
+- remove features: `layout`, `safe_color`, `safe_draw`, `safe_image`, `_destaque*`, `_stack*`.
 - rename features: `lin` to `alg`.
 - add more feature gates for: `time`, `ui`.
 - remove the `unsafe··` flag requirement for re-exported unsafe std methods.
@@ -65,6 +65,7 @@
 - new types: `Route`, `RouteAnchor`, `RouteName`, `RouteSeg`.
 
 #### data::layout
+- remove items: `Destaque`, `Stack`.
 - remove trait `DataDesta`.
 
 #### data::layout::array